@@ -16,17 +16,17 @@
     "ethers": "^5.5.4",
     "graphql-ws": "^5.15.0",
     "moment": "^2.30.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-router-dom": "^6.2.2",
+    "serve": "^11.3.2",
     "web3": "^1.7.1",
-    "web3modal": "^1.9.12",
-    "react-scripts": "^4.0.3",
-    "serve": "^11.3.2"
+    "web3modal": "^1.9.12"
   },
   "scripts": {
-    "start": "serve -s build",
+    "start": "react-scripts start",
     "build": "react-scripts build",
+    "serve": "serve -s build",
     "heroku-postbuild": "npm run build"
   },
   "eslintConfig": {
