@@ -1,16 +1,18 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App'
 import User from './contexts/User'
 import Mode from './contexts/Mode'
-import { Fonts } from './assets/fonts/fonts';
+import ErrorBoundary from './components/ErrorBoundary'
 
-ReactDOM.render(
-  <User>
-    <Mode>
-      <App />
-    </Mode>
-  </User>,
-  document.getElementById('root')
-)
+const root = createRoot(document.getElementById('root'));
+root.render(
+  <ErrorBoundary>
+    <User>
+      <Mode>
+        <App />
+      </Mode>
+    </User>
+  </ErrorBoundary>
+);
